README.md
P-tweets was formed for people to be able to chat with their friends, family and others on topics that they have the same interest in. P-tweets is named for Popular Tweeting on the social networking circuit.
It was formed in the summer of 2019 in the San Francisco Bay Area by a web developer and a host of programmers. Inspiration for this site came in early spring with the idea of developing a social blogging site. Not only was there an increasing desire to get people involved on the Internet with micro blogging, the Internet provided great opportunity to bring like minds together. The Internet also provided a network within a defined community of individuals with the same concerns.

I find most fascinating - exploring clusters of loosely related people by looking at the updates from their friends, family and others. There are stories told in between updates. Who’s at a conference, and do they know each other? Who’s on the road, and who’s at home. Narratives that wind around and between the updates and the people, that show connections. Updates that echo each other, or even directly respond to another P-tweets post.